Refactor for extensibility
* Versioning schemes are now strings instead of pre-defined enum values, making the implementation more flexible
* Users can now extend versatile with additional, or customized `Version` implementations via SPI
* Split the project into `versatile-core` and `versatile-spi` to facilitate SPI offering
* Schemes for which no `Version` implementation is found will fall back to `GenericVersion`, allowing versatile to function when working with versions from uncontrolled data sets
