+++ .github/release.yml
@@ -1,5 +1,8 @@
 changelog:
   categories:
+  - title: Breaking Changes 🚨
+    labels:
+    - breaking change
   - title: Enhancements 🚀
     labels:
     - enhancement
