--- src/main/java/io/github/nscuro/versatile/version/InvalidVersionException.java
@@ -1,33 +1,0 @@
-/*
- * This file is part of versatile.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * Copyright (c) Niklas Düster. All Rights Reserved.
- */
-package io.github.nscuro.versatile.version;
-
-import io.github.nscuro.versatile.VersException;
-
-public class InvalidVersionException extends VersException {
-
-    InvalidVersionException(final String version, final String message) {
-        this(version, message, null);
-    }
-
-    InvalidVersionException(final String version, final String message, final Throwable cause) {
-        super("Version %s is invalid: %s".formatted(version, message), cause);
-    }
-
-}
