+++ versatile-core/src/main/java/io/github/nscuro/versatile/Comparator.java
