+++ versatile-core/src/main/java/io/github/nscuro/versatile/PairwiseIterator.java
