+++ versatile-core/src/main/java/io/github/nscuro/versatile/VersException.java
