+++ versatile-core/src/main/java/io/github/nscuro/versatile/version/KnownVersioningSchemes.java
@@ -1,0 +1,36 @@
+/*
+ * This file is part of versatile.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright (c) Niklas Düster. All Rights Reserved.
+ */
+package io.github.nscuro.versatile.version;
+
+/**
+ * @since 0.8.0
+ */
+public final class KnownVersioningSchemes {
+
+    public static final String SCHEME_DEBIAN = "deb";
+    public static final String SCHEME_GENERIC = "generic";
+    public static final String SCHEME_GOLANG = "golang";
+    public static final String SCHEME_MAVEN = "maven";
+    public static final String SCHEME_NPM = "npm";
+    public static final String SCHEME_RPM = "rpm";
+
+    private KnownVersioningSchemes() {
+    }
+
+}
