+++ versatile-core/src/main/java/io/github/nscuro/versatile/version/ext/ComponentVersion.java
