+++ versatile-core/src/main/resources/META-INF/services/io.github.nscuro.versatile.version.VersionProvider
@@ -1,0 +1,6 @@
+io.github.nscuro.versatile.version.DebianVersionProvider
+io.github.nscuro.versatile.version.GenericVersionProvider
+io.github.nscuro.versatile.version.GoVersionProvider
+io.github.nscuro.versatile.version.MavenVersionProvider
+io.github.nscuro.versatile.version.NpmVersionProvider
+io.github.nscuro.versatile.version.RpmVersionProvider
