+++ versatile-core/src/test/java/io/github/nscuro/versatile/PairwiseIteratorTest.java
