+++ versatile-core/src/test/java/io/github/nscuro/versatile/version/RpmVersionTest.java
